EmployeeForm now handles new and edit requests
